financial_rag_chatbot/requirements.txt
# Web Framework
streamlit==1.33.0  # Latest stable version

# Vector Database
chromadb==0.4.24  
hnswlib==0.7.0

# Data Handling
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2

# PDF Processing
PyPDF2==3.0.1

sentence-transformers==2.2.2


# NLP and Embeddings
spacy==3.7.2
spacy-transformers==1.3.4  # Needed for transformer-based embeddings

# Machine Learning & Deep Learning
torch==2.3.0
torchaudio==2.3.0

# FAISS for Vector Search
faiss-cpu==1.7.4  # Ensures compatibility with NumPy



# Avoid OpenMP Conflicts
nomkl  # Disables MKL to prevent OpenMP issues
mkl-service==2.4.0  # Ensures proper MKL linkage
scipy==1.13.0  # Latest stable version for compatibility
